rename basic.rs -> type_enum.rs
